Revert "while testing"
This reverts commit 6ec5dbc60824ab59697ff1f40e7b0520a9e42697.

+++ src/main/java/DeliveryV1/CoordinateVisualizer.java
@@ -76,7 +76,7 @@ public class CoordinateVisualizer extends JPanel {
                 g.drawLine(startX, startY, centerX, centerY);
                 if (route.totalDistance != 0)
                 {
-                    g.drawString("Total Distance: " + (int) route.totalDistance, startX, startY);
+                    g.drawString("Total Distance: " + (double) route.totalDistance, startX, startY);
                 }
             }
         }
